refactor: merge CPDAGSHD into SHD for auto PDAG detection
